codigo refatorado
