+++ ds2m/caio_palermo_lemos/js/convert-timezone.js
@@ -1,0 +1,7 @@
+export function convertTZ(date, tzString) {
+  return new Date(
+    (typeof date === "string" ? new Date(date) : date).toLocaleString("en-US", {
+      timeZone: tzString,
+    })
+  );
+}
